+++ AdaptixServer/core/connector/connector.go
@@ -67,7 +67,7 @@ type Teamserver interface {
 
 	TsChatSendMessage(username string, message string)
 
-	TsDownloadAdd(agentId string, fileId string, fileName string, fileSize int) error
+	TsDownloadAdd(agentId string, fileId string, fileName string, fileSize int64) error
 	TsDownloadUpdate(fileId string, state int, data []byte) error
 	TsDownloadClose(fileId string, reason int) error
